requirements.txt
Django==2.2.6
djangorestframework==3.10.3
flake8==3.7.8
psycopg2>=2.7.5,<2.8.0